requirements.txt
torch
torchvision
diffusers
pillow
accelerate
matplotlib
safetensors
wandb
mosaicml
